add remove user action

+++ internal/entities/user.go
@@ -8,8 +8,8 @@ type User struct {
 type UserRepository interface {
 	Create(user *User) error
 	Update(user *User) error
-	Delete(userId int64) error
-	GetById(userId int64) (User, error)
+	Delete(userId int32) error
+	GetById(userId int32) (User, error)
 	GetByName(name string) (User, error)
 	Fetch(num int64) ([]User, error)
